feat: Add SearchItem click event
